Update interfaceDENTRO.txt
Correction of Dutch translations.
Remark 1: in the original file, sometimes a description consists of several words connected by dots and sometimes with a space between the words. I kept things as they were, but this is not always consistent.
Remark 2: last responsetrait (16): in our original database, this was scored as "no big risks of diseases", but here it is described as "diseases", which is the opposite... To be checked. Our idea was that all scores should be "positive".