For now, DFinfo calculated here. Later will try for other models - make it totaly general and implement in different place.
Now the function is not hard-coded, it will load all "info" columns in data - if they incluced TRUE/FALSE - it will rename it by choice - so it can be later easilly translated. If the values are custom it will just use them. It will also agregate "info" columns of same "criteria" - eg, all food or wood columns are aggregated for each tree.